// src/common/error/handle-errors.enum.ts
export enum ErrorCodeEnum {
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',
  CODE_EXPIRED = 'CODE_EXPIRED',
  INVALID_CODE = 'INVALID_CODE',
  PASSWORD_REQUIRED = 'PASSWORD_REQUIRED',
  SAME_PASSWORD = 'SAME_PASSWORD',
  INVALID_TAG = 'INVALID_TAG',
  EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
  GOOGLE_MISSING_REQUIRED_FIELDS = 'GOOGLE_MISSING_REQUIRED_FIELDS',
}